split EditHeader and ViewHeader from Header
Split out the EditHeader and ViewHeader into separate component modules.
